requirements_test.txt
# Testing requirements
coverage==4.3.3
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
django-model-utils>=2.0

# Additional app
django-annoying>=0.10.3
